% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSep-package.r
\docType{package}
\name{dSep-package}
\alias{dSep-package}
\title{dSep-package}
\details{
\packageDESCRIPTION{dSep}
\packageIndices{dSep}
}
\references{
Gonzalez-Voyer, Alejandro, and Achaz Von Hardenberg. 2014. "An Introduction to Phylogenetic Path Analysis." In Modern Phylogenetic Comparative Methods and Their Application in Evolutionary Biology, edited by Laszlo Zsolt Garamszegi, 29. Berlin, Heidelberg: Springer Berlin Heidelberg.
\url{http://www.mpcm-evolution.org/practice/online-practical-material-chapter-8/chapter-8-2-step-step-guide-phylogenetic-path-analysis-using-d-sep-method-rhinograds-example}
}
\author{
Joan Maspons <\email{j.maspons@creaf.uab.cat}>
}
