% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D-separation.R, R/plot.R
\name{dSep}
\alias{dSep}
\alias{plot.dSep}
\title{Path analysis usding d-separation}
\usage{
\method{plot}{dSep}(
  x,
  y,
  plotCoef = TRUE,
  plotdSep = FALSE,
  legend = TRUE,
  lty = c(signif = 1, nonSignif = 2),
  color = c(causal = "black", dsep = "red"),
  alpha = 0.05,
  ...
)
}
\arguments{
\item{x}{a dSep object.}

\item{plotCoef}{if \code{TRUE} plot the path coefficients.}

\item{plotdSep}{if \code{TRUE} plot the d-separation paths. By default in red.}

\item{legend}{if \code{TRUE} add a legend.}

\item{lty}{a vector of 2 elements with the line type of significant and non significant paths.}

\item{color}{a vector of 2 elements with the color of the causal paths and the color of the d-separation paths.}

\item{alpha}{significance level.}

\item{...}{parameters passed to FUN. Parameters must be named following the FUN arguments (e.g. data=data.frame()).}

\item{FUN}{a function or a the name of the function to test the conditional independences.
Currently tested with \code{\link[stats]{lm}}, \code{\link[stats]{glm}}, \code{\link[nlme]{gls}}, \code{\link[caper]{pgls}}, \code{\link[MCMCglmm]{MCMCglmm}} and \code{\link[brms]{brm}}.}

\item{formulaArg}{argument name from FUN that accepts the formula parameter.}

\item{orderResponse}{parameter passed to \code{\link{condIndep}}.}

\item{n}{sample size of the dataset. If omited the function tries to guess from the models using \code{\link{nobs}} and custom functions.}

\item{cl}{the number of CPU cores or a cluster object to run the models in parallel. Cluster object can be defined with \code{\link[parallel]{makeCluster}}
in package \code{parallel} or \code{\link[snow]{makeCluster}} from \code{snow} package.}

\item{pathCoef}{if \code{TRUE} calculates the path coefficient with \code{\link{pathCoef}}. It is advised to standardize the
input data (see \code{\link{scale}}). By doing this, the standardized coefficients represents the relative
strength of each causal relationship in the model.}
}
\value{
a dSep object.
}
\description{
Path analysis usding d-separation
}
\examples{
## Dummy data
g1<- gRbase::dag(~a:c:d + b:d)
g2<- gRbase::dag(~a:c:d + b:d:a)
g<- list(m1=g1, m2=g2)
d<- data.frame(a=rnorm(100), b=rnorm(100), c=rnorm(100), d=rnorm(100))

## Use of different functions
m.lm<- dSep(g1, FUN="lm", n=nrow(d), data=d)
m.glm<- dSep(g, FUN="glm", n=nrow(d), data=d)

if (require("nlme"))
  m.gls<- dSep(g1, FUN=gls, formulaArg="model", n=nrow(d), data=d)

if (require(caper)){
  data(shorebird, package="caper")
  shorebird.data[,2:5]<- scale(log(shorebird.data[,2:5]))
  shorebird<- comparative.data(shorebird.tree, shorebird.data, 'Species')
  g3<- gRbase::dag(~Egg.Mass:M.Mass:F.Mass)
  g4<- gRbase::dag(~Egg.Mass:M.Mass:Cl.size + Cl.size:F.Mass:M.Mass)
  m1.pgls<- dSep(list(mPhy1=g3, mPhy2=g4), FUN=caper::pgls, n=nrow(d), cl=2, data=shorebird, lambda='ML')

  rhino.dat <- read.csv("http://mpcm-evolution.org/OPM/Chapter8_OPM/download/rhino.csv")
  rhino.tree <- read.tree("http://mpcm-evolution.org/OPM/Chapter8_OPM/download/rhino.tree")
  com.dat<- caper::comparative.data(rhino.tree, rhino.dat, SP, vcv=TRUE, vcv.dim=3, warn.dropped=TRUE)
  m<- list()
  m$h1<- gRbase::dag(~LS:BM + NL:BM + DD:NL + RS:DD)
  m$h2<- gRbase::dag(~LS:BM + NL:BM + DD:NL + RS:DD:LS)
  m$h3<- gRbase::dag(~LS:BM + NL:BM + DD:NL + RS:NL)
  m$h4<- gRbase::dag(~LS:BM + NL:BM + DD:NL + RS:BM:NL)
  m$h5<- gRbase::dag(~LS:BM + NL:BM + DD:NL + RS:BM:NL:DD)
  m$h6<- gRbase::dag(~LS:BM + NL:BM:RS + DD:NL + RS:BM)
  m$h7<- gRbase::dag(~LS:BM + NL:BM:RS + DD:NL + RS:BM:LS)
  m$h8<- gRbase::dag(~LS:BM + NL:BM:RS + DD:NL)
  m$h9<- gRbase::dag(~LS:BM + NL:BM:RS + DD:NL + RS:LS)

  m2.pgls<- dSep(m, FUN=pgls, cl=2, orderResponse=c("BM", "RS", "LS", "NL", "DD"), data=com.dat, lambda="ML")
}

if (require("brms"))
  m.brmfit<- dSep(g1, FUN=brm, n=nrow(d), cl=2, pathCoef=FALSE, data=d)

if (require("MCMCglmm"))
  m.MCMCglmm<- dSep(g1, FUN=MCMCglmm, formulaArg="fixed", n=nrow(d), cl=2, data=d, verbose=FALSE)

##
plot(m.glm)
}
\references{
Gonzalez-Voyer, Alejandro, and Achaz Von Hardenberg. 2014. "An Introduction to Phylogenetic Path Analysis." In Modern Phylogenetic Comparative Methods and Their Application in Evolutionary Biology, edited by Laszlo Zsolt Garamszegi, 29. Berlin, Heidelberg: Springer Berlin Heidelberg.
\url{http://www.mpcm-evolution.org/practice/online-practical-material-chapter-8/chapter-8-2-step-step-guide-phylogenetic-path-analysis-using-d-sep-method-rhinograds-example}
}
\author{
Joan Maspons <\email{j.maspons@creaf.uab.cat}>
}
