% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{edgeAttrs}
\alias{edgeAttrs}
\title{Edge attributes for ploting dSep and pathCoef classes}
\usage{
edgeAttrs(g, alpha = 0.05, lty = 1:2, color = 1:2, ...)
}
\arguments{
\item{g}{a \linkS4class{graph}.}

\item{alpha}{significance level.}

\item{lty}{a vector of 2 elements with the line type of significant and non significant paths.}

\item{color}{a vector of 2 elements with the color of the causal edges and the color of the d-separation edges.}

\item{...}{other valid edge attributes.}
}
\description{
Edge attributes for ploting dSep and pathCoef classes
}
