% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{merge.graph}
\alias{merge.graph}
\title{Merge graphs}
\usage{
\method{merge}{graph}(x, y, ...)
}
\arguments{
\item{x}{\linkS4class{graph}}

\item{y}{\linkS4class{graph}}
}
\description{
Merge graphs
}
\details{
If the same edge attribute is present in both, the value from the second one is omited.
TODO: merge nodeData and graphData
}
