% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathCoef.R, R/plot.R
\name{pathCoef}
\alias{pathCoef}
\alias{pathCoef.dSep}
\alias{pathCoef.graph}
\alias{pathCoef.list}
\alias{summary.pathCoef}
\alias{plot.pathCoef}
\title{Path coefficients
It is advised to standardize the input data (see \code{\link{scale}}). By doing this, the standardized coefficients represents the relative
strength of each causal relationship in the model.}
\usage{
pathCoef(x, FUN = "lm", formulaArg = "formula", cl, alpha = 0.05, ...)

\method{pathCoef}{dSep}(x, ...)

\method{pathCoef}{graph}(x, ...)

\method{pathCoef}{list}(x, FUN = "lm", formulaArg = "formula", cl, alpha = 0.05, ...)

\method{summary}{pathCoef}(object, ...)

\method{plot}{pathCoef}(
  x,
  y,
  plotCoef = TRUE,
  legend = TRUE,
  lty = c(signif = 1, nonSignif = 2),
  alpha = 0.05,
  ...
)
}
\arguments{
\item{x}{a \code{pathCoef} object.}

\item{FUN}{a function or a the name of the function to test the conditional independences.
Currently tested with \code{\link[stats]{lm}}, \code{\link[stats]{glm}}, \code{\link[nlme]{gls}}, \code{\link[caper]{pgls}}, \code{\link[MCMCglmm]{MCMCglmm}} and \code{\link[brms]{brm}}.}

\item{formulaArg}{argument name from FUN that accepts the formula parameter of the lineal model.}

\item{cl}{the number of CPU cores or a cluster object to run the models in parallel. Cluster object can be defined with \code{\link[parallel]{makeCluster}}
in package \code{parallel} or \code{\link[snow]{makeCluster}} from \code{snow} package.}

\item{alpha}{significance level.}

\item{...}{parameters passed to FUN. Parameters must be named following the FUN arguments (e.g. data=data.frame()).}

\item{plotCoef}{if \code{TRUE} plot the path coefficients.}

\item{legend}{if \code{TRUE} add a legend.}

\item{lty}{a vector of 2 elements with the line type of significant and non significant paths.}
}
\value{
a pathCoef object.
}
\description{
Path coefficients
It is advised to standardize the input data (see \code{\link{scale}}). By doing this, the standardized coefficients represents the relative
strength of each causal relationship in the model.
}
\examples{
## Dummy data
g1<- gRbase::dag(~a:c:d + b:d)
g2<- gRbase::dag(~a:c:d + b:d:a)
g<- list(m1=g1, m2=g2)
d<- data.frame(a=rnorm(100), b=rnorm(100), c=rnorm(100), d=rnorm(100))

p.a<- pathCoef(g1, FUN="lm", nobs=nrow(d), data=d)
plot(p.b<- pathCoef(list(m1=g1, m2=g2), FUN="lm", nobs=nrow(d), data=d))
}
\author{
Joan Maspons <\email{j.maspons@creaf.uab.cat}>
}
