% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM-methods.R
\name{pValues}
\alias{pValues}
\alias{pValues.default}
\alias{pValues.lm}
\alias{pValues.pgls}
\alias{pValues.phylolm}
\alias{pValues.phyloglm}
\alias{pValues.MCMCglmm}
\alias{pValues.gls}
\alias{pValues.brmsfit}
\alias{pValues.lme}
\title{Exctract p-value of the predictors}
\usage{
pValues(x)

\method{pValues}{default}(x)

\method{pValues}{lm}(x)

\method{pValues}{pgls}(x)

\method{pValues}{phylolm}(x)

\method{pValues}{phyloglm}(x)

\method{pValues}{MCMCglmm}(x)

\method{pValues}{gls}(x)

\method{pValues}{brmsfit}(x)

\method{pValues}{lme}(x)
}
\arguments{
\item{x}{the result of a model from lm, glm, gls, pgls, MCMCglmm or brm functions}
}
\value{
A named vector with p-values for each predictor
}
\description{
Exctract p-value of the predictors
}
